# Conflicts:
#	src/main/java/com/cydeo/javahedgehogsproject/controller/SalesInvoiceController.java
#	src/main/java/com/cydeo/javahedgehogsproject/service/InvoiceProductService.java
#	src/main/java/com/cydeo/javahedgehogsproject/service/implementation/InvoiceProductServiceImpl.java

+++ src/main/java/com/cydeo/javahedgehogsproject/entity/InvoiceProduct.java
@@ -8,6 +8,7 @@ import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -17,7 +18,7 @@ public class InvoiceProduct extends  BaseEntity{
 
     private int quantity;
     private BigDecimal price;
-    private int tax;
+    private BigDecimal tax;
     private BigDecimal profitLoss;
     private int remainingQty;
     @ManyToOne
